chore: Move `shiny add test` functions to `_main_add_test.py`
